vision/requirements.txt
fastapi==0.135.3
httpx==0.28.1
opencv-python-headless==4.13.0.92
pydantic==2.13.1
python-multipart==0.0.26
torch==2.11.0+cpu
torchvision==0.26.0+cpu
ultralytics==8.4.37
uvicorn==0.44.0
timm==1.0.26
git+https://github.com/ultralytics/CLIP.git
open_clip_torch==2.30.0
